Vovila/gardening seeds maker storages

+++ Assets/Modules/Gardening/HarvestResult.cs
@@ -1,11 +1,9 @@
-using Modules.Products.Plants;
-
 namespace Modules.Gardening
 {
     public struct HarvestResult
     {
         public bool IsCollected;
         public PlantType Type;
-        public float Value;
+        public int Value;
     }
 }
